+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.1] - 2022-01-25
 - Initial repo release
 - Added graphql fetchers and types
+- Update queries to handle pagination and network info by default as needed