+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.2.0
+
+### Minor Changes
+
+- 77c330d: Moving types to types.ts and zdk to zdk.ts, adding offchain and nouns queries
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -8,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ## [0.0.1] - 2022-01-25
+
 - Initial repo release
 - Added graphql fetchers and types
 - Update queries to handle pagination and network info by default as needed